destructor must be virtual for exception, while we are using inheritance

+++ common/base_cpp/exception.cpp
@@ -38,7 +38,7 @@ void Exception::appendMessage(const char* format, ...)
     va_list args;
     va_start(args, format);
 
-    const int len = strlen(_message);
+    const size_t len = static_cast<size_t>(strlen(_message));
     vsnprintf(_message + len, sizeof(_message) - len, format, args);
 
     va_end(args);
